Reduce hello example cmd to one runner, clean up
